+++ mono/checklists/serializers.py
@@ -5,7 +5,7 @@ from rest_framework.serializers import (
     CurrentUserDefault, HiddenField, ModelSerializer, Serializer,
 )
 
-from .models import Checklist, Task
+from .models import Checklist, Configuration, Task
 
 
 class ChecklistSerializer(ModelSerializer):
@@ -56,8 +56,16 @@ class TaskSerializer(ModelSerializer):
             'reminder',
             'reminded',
             'due_date',
+            'recurrence',
+            'get_recurrence_display',
+            'next_task_created',
         ]
-        read_only_fields = ['created_by', 'checked_by', 'checked_at']
+        read_only_fields = [
+            'created_by',
+            'checked_by',
+            'checked_at',
+            'get_recurrence_display',
+        ]
 
     @transaction.atomic
     def create(self, validated_data):
@@ -153,3 +161,24 @@ class TaskMoveSerializer(Serializer):
         )
         order = self.validated_data['order']
         task.set_order(order)
+
+
+class ConfigurationSerializer(serializers.ModelSerializer):
+    """Configuration serializer"""
+    created_by = serializers.HiddenField(
+        default=serializers.CurrentUserDefault()
+    )
+
+    class Meta:
+        model = Configuration
+        fields = [
+            'id',
+            'show_completed_tasks',
+            'created_at',
+            'updated_at',
+            'created_by',
+        ]
+
+
+class TaskRecurrenceSerializer(serializers.Serializer):
+    recurrence = 1