+++ mono/checklists/static/checklists/index.js
@@ -1,14 +1,15 @@
-function toast(title, message) {
+function toast(title, message, type) {
     $('body').toast({
         title: title,
-        message: message
+        message: message,
+        class: type ? type: '',
     })
 }
 
 function renderChecklist(checklistId, checklistName) {
     listsDiv.append(`
         <a class="teal checklist item" data-checklist-id="${checklistId}" onclick="selectChecklist(${checklistId})" style="display: flex; flex-flow: row nowrap; align-items: center;">
-            <div style="flex: 1 0 0;">
+            <div style="flex: 1 0 0; white-space: nowrap; overflow: hidden; text-overflow: ellipsis;">
                 <span class="checklist-name">${checklistName}</span>
             </div>
             <div style="flex: 0 0 auto;">
@@ -34,7 +35,7 @@ function renderTask(task) {
             <div class="ui checkbox" data-task-id="${task.id}" style="flex: 0 0 auto;">
                 <input type="checkbox" ${task.checked_at != null ? "checked" : ""}>
             </div>
-            <div style="flex: 1 0 auto;">
+            <div style="flex: 1 0 0; white-space: nowrap; overflow: hidden; text-overflow: ellipsis;">
                 <span class="task-description">${task.description}</span>
             </div>
             ${task.reminder ? `
@@ -45,6 +46,11 @@ function renderTask(task) {
             ${task.due_date ? `
                 <div class="ui circular icon ${due ? 'red': ''} label task-label" style="margin-left: .5em;" title="${stringToLocalDate(task.due_date, languageCode)}">
                     <i class="calendar icon"></i>
+                </div>
+            `: ''}
+            ${task.recurrence ? `
+                <div class="ui circular icon ${task.next_task_created ? '': 'blue'} label task-label" style="margin-left: .5em;" title="${task.get_recurrence_display}">
+                    <i class="retweet icon"></i>
                 </div>
             `: ''}
         </a>
@@ -85,11 +91,19 @@ function createTask(taskDescription) {
         data: {
             checklist: checklistId,
             description: taskDescription,
-            order: 5
+            order: 0
         },
-        onSuccess: r => {
+        onSuccess(r) {
             renderTask(r)
-        }
+            updateCounter('total', true)
+        },
+        onFailure(response, element, xhr) {
+            Object.entries(response).forEach(
+                entry => {
+                    toast(gettext('Error'), entry[1].join('\n'), 'error')
+                }
+            )
+        },
     })
 }
 
@@ -101,9 +115,16 @@ async function createChecklist(checklistName) {
         data: {
             name: checklistName,
         },
-        onSuccess: r => {
+        onSuccess(r) {
             renderChecklist(r.id, r.name)
-        }
+        },
+        onFailure(response, element, xhr) {
+            Object.entries(response).forEach(
+                entry => {
+                    toast(gettext('Error'), entry[1].join('\n'), 'error')
+                }
+            )
+        },
     })
 }
 
@@ -112,7 +133,7 @@ function retrieveLists() {
         on: 'now',
         url: '/cl/api/checklists/',
         successTest: r => true,
-        onSuccess: r => {
+        onSuccess(r) {
             listsDiv.empty()
             r.results.forEach(
                 item => renderChecklist(item.id, item.name)
@@ -127,27 +148,47 @@ function selectChecklist(checklistId) {
     hideTaskPanel()
     $('*.checklist.item').removeClass('active')
     $(`.checklist.item[data-checklist-id=${checklistId}]`).addClass('active')
+    initializeConfigMenu(false)
+}
+
+function reselectChecklist() {
+    selectedChecklist = sessionStorage.getItem('selectedChecklist')
+    if (selectedChecklist !== null) {
+        selectChecklist(selectedChecklist)
+    }
+}
+
+async function updateCounter(type, add) {
+    if (type == 'total') {
+        el = $('#total-tasks')
+    } else {
+        el = $('#completed-tasks')
+    }
+    if (add) {
+        el.text(parseInt(el.text()) + 1)
+    } else {
+        el.text(parseInt(el.text()) - 1)
+    }
 }
 
 async function retrieveTasks(checklistId) {
     sessionStorage.setItem('selectedChecklist', checklistId)
     $.api({
-        url: `/cl/api/checklists/${checklistId}/`,
-        on: 'now',
-        method: 'GET',
-        stateContext: '#tasks-div',
-        onSuccess: checklist => {
-            $('#list-name').text(checklist.name);
-        }
-    })
-    $.api({
         url: `/cl/api/tasks/?checklist__id=${checklistId}`,
         on: 'now',
         method: 'GET',
         stateContext: '#tasks-div',
-        onSuccess: r => {
-            tasksDiv.empty();
-            r.results.forEach(
+        onSuccess(r) {
+            tasksDiv.empty()
+            $('#counter .label').removeClass('hidden')
+            $('#completed-tasks').text(r.results.filter(r => r.checked_at != null).length)
+            $('#total-tasks').text(r.results.length)
+            if (JSON.parse(sessionStorage.getItem('showCompletedTasks')) === true) {
+                tasks = r.results
+            } else {
+                tasks = r.results.filter(t => t.checked_at === null)
+            }
+            tasks.forEach(
                 item => {
                     renderTask(item)
                     selectedTask = sessionStorage.getItem('selectedTask')
@@ -157,6 +198,13 @@ async function retrieveTasks(checklistId) {
             initializeTaskList()
         }
     })
+}
+
+function reselectTask() {
+    selectedTask = sessionStorage.getItem('selectedTask')
+    if (selectedTask !== null) {
+        selectTask(selectedTask)
+    }
 }
 
 function checkTask(taskId) {
@@ -171,6 +219,13 @@ function checkTask(taskId) {
             if (taskId === selectedTask) {
                 $('#check-icon').addClass('check circle outline')
                 $('#task-description').attr('data-checked', true)
+            }
+            updateCounter('completed', true)
+            if (JSON.parse(sessionStorage.getItem('showCompletedTasks')) === true) {
+                reselectTask()
+            } else {
+                sessionStorage.removeItem('selectedTask')
+                reselectChecklist()
             }
         }
     })
@@ -189,6 +244,8 @@ function uncheckTask(taskId) {
                 $('#check-icon').removeClass('check circle outline').addClass('circle outline')
                 $('#task-description').attr('data-checked', false)
             }
+            reselectTask()
+            updateCounter('completed', false)
         }
     })
 }
@@ -214,10 +271,15 @@ function updateTask(data) {
         data: data,
         onSuccess(response, element, xhr) {
             retrieveTasks(checklistId)
+            reselectTask()
         },
         onFailure(response, element, xhr) {
-            toast(response)
-        }
+            Object.entries(response).forEach(
+                entry => {
+                    toast(gettext('Error'), entry[1].join('\n'), 'error')
+                }
+            )
+        },
     })
 }
 
@@ -228,7 +290,7 @@ function selectTask(taskId) {
         on: 'now',
         url: `/cl/api/tasks/${taskId}/`,
         stateContext: '#task-detail .segment',
-        onSuccess: r => {
+        onSuccess(r) {
             sessionStorage.setItem('selectedTask', taskId)
             renderTaskDetails(r)
             showTaskPanel()
@@ -265,6 +327,13 @@ function renderTaskDetails(task) {
         localDueDate = new Date(task.due_date);
         utcDueDate = new Date(localDueDate.toUTCString().slice(0, -4));
         $('#task-due-date').calendar('set date', utcDueDate, true, false)
+    }
+    if (task.recurrence === null) {
+        $('#task-recurrence').dropdown('clear', true)
+        $('#task-recurrence').removeClass('blue button').addClass('button')
+    } else {
+        $('#task-recurrence').dropdown('set selected', task.recurrence, undefined, true)
+        $('#task-recurrence').removeClass('button').addClass('blue button')
     }
 }
 
@@ -306,14 +375,14 @@ function deleteTask() {
                 class: 'red approve'
             },
         ],
-        onApprove: () => {
+        onApprove() {
             taskId = sessionStorage.getItem('selectedTask')
             checklistId = sessionStorage.getItem('selectedChecklist')
             $.api({
                 on: 'now',
                 method: 'DELETE',
                 url: `/cl/api/tasks/${taskId}`,
-                onSuccess: r => {
+                onSuccess(r) {
                     sessionStorage.removeItem('selectedTask')
                     retrieveTasks(checklistId)
                     hideTaskPanel()
@@ -322,6 +391,7 @@ function deleteTask() {
         }
     }).modal('show');
 }
+
 function deleteChecklist(id) {
     $('.ui.sidebar').sidebar('hide')
     $('body').modal({
@@ -344,7 +414,7 @@ function deleteChecklist(id) {
                 on: 'now',
                 method: 'DELETE',
                 url: `/cl/api/checklists/${id}`,
-                onSuccess: r => {
+                onSuccess(r) {
                     if (id == sessionStorage.getItem('selectedChecklist')) {
                         sessionStorage.removeItem('selectedChecklist')
                         hideTaskPanel()
@@ -361,16 +431,16 @@ function showNewChecklistModal() {
     modal = $('#new-checklist')
     input = modal.find('input[name=name]')
     modal.modal({
-        onShow: () => {
+        onShow() {
             input.val('')
-            input.off().on('keyup', function (e) {
-                if (e.key === 'Enter' || e.keyCode === 13) {
+            input.off().on('keypress', function (e) {
+                if (e.keycode === 13 || e.which === 13) {
                     createChecklist(input.val())
                     modal.modal('hide')
                 }
             })
         },
-        onApprove: () => {
+        onApprove() {
             createChecklist(input.val())
         }
     }).modal('show')
@@ -382,16 +452,16 @@ function editChecklist(checklistId) {
     input = modal.find('input[name=name]')
     selectedChecklist = sessionStorage.getItem('selectedChecklist')
     modal.modal({
-        onShow: () => {
+        onShow() {
             input.val($(`.checklist.item[data-checklist-id=${checklistId}] .checklist-name`).first().text())
-            input.off().on('keyup', function (e) {
-                if (e.key === 'Enter' || e.keyCode === 13) {
+            input.off().on('keypress', function (e) {
+                if (e.keycode === 13 || e.which === 13) {
                     renameChecklist(checklistId, input.val())
                     modal.modal('hide')
                 }
             })
         },
-        onApprove: () => {
+        onApprove() {
             renameChecklist(checklistId, input.val())
         }
     }).modal('show')
@@ -414,16 +484,16 @@ function showNewTaskModal() {
     modal = $('#new-task')
     input = modal.find('input[name=name]')
     modal.modal({
-        onShow: () => {
+        onShow() {
             input.val('')
-            input.off().on('keyup', function (e) {
-                if (e.key === 'Enter' || e.keyCode === 13) {
+            input.off().on('keypress', function (e) {
+                if (e.keycode === 13 || e.which === 13) {
                     createTask(input.val())
                     modal.modal('hide')
                 }
             })
         },
-        onApprove: () => {
+        onApprove() {
             createTask(input.val())
         }
     }).modal('show')
@@ -466,7 +536,6 @@ function initializeDragAndDrop() {
             on: 'now',
             method: 'POST',
             url: '/cl/api/task-move/',
-            headers: { 'X-CSRFToken': csrftoken },
             stateContext: '#tasks-div',
             data: {
                 task: taskId,
@@ -486,7 +555,6 @@ function initializeDragAndDrop() {
             on: 'now',
             method: 'POST',
             url: '/cl/api/checklist-move/',
-            headers: { 'X-CSRFToken': csrftoken },
             stateContext: '#lists-div',
             data: {
                 checklist: checlistId,
@@ -499,13 +567,17 @@ function initializeDragAndDrop() {
 
 function initializeTaskList() {
     clearSearch()
-    var options = {
-        listClass: 'filter-list',
-        valueNames: [
-            'task-description',
-        ]
-    };
-    taskList = new List('task-list-panel', options);
+    if ($('.task.item').length > 0) {
+        taskList = new List(
+            'task-list-panel',
+            {
+                listClass: 'filter-list',
+                valueNames: [
+                    'task-description',
+                ]
+            },
+        )
+    }
 }
 
 function clearSearch() {
@@ -519,7 +591,9 @@ function initializeDueDate() {
         onChange() {
             dueDate = $('#task-due-date').calendar('get date')
             if (dueDate != null) {
-                dueDate = $('#task-due-date').calendar('get date').toISOString().split('T')[0]
+                dueDate = $('#task-due-date').calendar('get date')
+                dueDate.setHours(0,0,0,0)
+                dueDate = dueDate.toISOString()
             }
             updateTask({
                 due_date: dueDate
@@ -527,8 +601,6 @@ function initializeDueDate() {
         }
     })
 }
-
-
 
 function initializeReminder() {
     $('#task-reminder').calendar({
@@ -551,7 +623,9 @@ function initializeReminder() {
 function clearDueDate() {
     $('#task-due-date').calendar('set date', null, true, false)
     updateTask({
-        due_date: null
+        due_date: null,
+        recurrence: null,
+        next_task_created: false,
     })
 }
 
@@ -560,5 +634,147 @@ function clearReminder() {
     updateTask({
         reminder: null,
         reminded: false,
+    })
+}
+
+function initializeConfigMenu(triggerReselect = true) {
+    selectedChecklist = sessionStorage.getItem('selectedChecklist')
+    if (selectedChecklist !== null) {
+        values = [
+            {
+                name: gettext('Sort by'),
+                type: 'left menu',
+                values: [
+                    {
+                        name: gettext('Description'),
+                        type: 'left menu',
+                        values: [
+                            {
+                                name: gettext('A ⭢ Z'),
+                                icon: 'sort alphabet down',
+                                value: 'sort-by-description-asc',
+                            },
+                            {
+                                name: gettext('Z ⭢ A'),
+                                icon: 'sort alphabet up',
+                                value: 'sort-by-description-desc',
+                            },
+                        ],
+                    },
+                    {
+                        name: gettext('Creation date'),
+                        type: 'left menu',
+                        values: [
+                            {
+                                name: gettext('Oldest first'),
+                                icon: 'sort amount down',
+                                value: 'sort-by-created_at-asc',
+                            },
+                            {
+                                name: gettext('Newest first'),
+                                icon: 'sort amount up',
+                                value: 'sort-by-created_at-desc',
+                            },
+                        ],
+                    },
+                ]
+            },
+            
+        ]
+    } else {
+        values = []
+    }
+
+    $.api({
+        on: 'now',
+        method: 'GET',
+        url: '/cl/api/config/',
+        onSuccess(r) {
+            sessionStorage.setItem('showCompletedTasks', r.show_completed_tasks)
+            if (triggerReselect) reselectChecklist()
+            if (r.show_completed_tasks == true) {
+                values.unshift(
+                    {
+                        name: gettext('Hide completed tasks'),
+                        icon: 'eye slash',
+                        value: 'hide-completed-tasks',
+                    }
+                )
+            } else {
+                values.unshift(
+                    {
+                        name: gettext('Show completed tasks'),
+                        icon: 'eye',
+                        value: 'show-completed-tasks',
+                    }
+                )
+            }
+            $('#config-menu').dropdown({
+                values: values,
+                action: 'hide',
+                onChange(value, text, $choice) {
+                    switch (true) {
+                        case value == 'show-completed-tasks':
+                            $.api({
+                                on: 'now',
+                                method: 'PUT',
+                                url: '/cl/api/config/',
+                                data: { show_completed_tasks: true },
+                                onSuccess(r) {
+                                    initializeConfigMenu()
+                                }
+                            })
+                            break
+                        case value =='hide-completed-tasks':
+                            $.api({
+                                on: 'now',
+                                method: 'PUT',
+                                url: '/cl/api/config/',
+                                data: { show_completed_tasks: false },
+                                onSuccess(r) {
+                                    initializeConfigMenu()
+                                }
+                            })
+                            break
+                        case value.startsWith('sort-by-'):
+                            selectedChecklist = sessionStorage.getItem('selectedChecklist')
+                            if (selectedChecklist === null) return
+                            field = value.replace('sort-by-', '').split('-')[0]
+                            direction = value.replace('sort-by-', '').split('-')[1]
+                            $.api({
+                                on: 'now',
+                                method: 'POST',
+                                url: `/cl/api/checklists/${selectedChecklist}/sort/`,
+                                data: {
+                                    field: field,
+                                    direction: direction,
+                                },
+                                onSuccess(r) {
+                                    reselectChecklist()
+                                }
+                            })
+                            break
+                    }
+                }
+            })
+        }
+    })
+}
+
+function initializeRecurrence() {
+    el = $('#task-recurrence')
+    el.dropdown({
+        values: taskRecurrenceValues,
+        onChange(value, text, $choice) {
+            data = { recurrence: value }
+            dueDate = $('#task-due-date').calendar('get date')
+            if (dueDate == null) {
+                dueDate = new Date()
+                dueDate.setHours(0,0,0,0)
+                dueDate = dueDate.toISOString()
+                data['due_date'] = dueDate
+            }
+            updateTask(data)
+        }
     })
 }